requirements.txt
python-telegram-bot
Pillow
werkzeug
